requirement.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
langchain-openai
sentence_transformers==2.2.2
InstructorEmbedding